==> code/analysis/02c_GRCCA-transcripts/03_characterize-GRCCA-transcript-res-GSEA.qmd <==
---
title: "02c-03. Characterize transcript-level GRCCA results"
author: "Rachel Smith"
format: html
editor: visual
---

Assess functional, cell-type, and risk gene enrichments of transcript-level GRCCA results.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
base_dir <- "~/Documents/PhD/projects/sgacc_wgcna_grcca/"
```

## Setup

```{r load_data}
source(paste0(base_dir, "scripts/load_WTCNA_res.R"))

## Load GRCCA data
load(paste0(objects_dir, "GRCCA_transcript_results.Rdata")) # df_lvs, df_y_res, df_x_res
```

## Get ranked gene list

```{r ranked_list}
my_grcca_ranked <- df_x_res_transcripts %>%
    group_by(ensembl_gene_id) %>%
    summarise(pearsons_r = median(pearsons_r)) %>%
    arrange(-pearsons_r) %>%
    dplyr::select(ensembl_gene_id, pearsons_r) %>%
    deframe()
```

## Cell-type GSEA

```{r cell_type_gsea}
df_grcca_fgsea_cell <- fgsea(
    pathways = cell_types,
    stats = my_grcca_ranked,
    eps = 0
) %>%
    as_tibble() %>%
    dplyr::rename("cell_type" = "pathway") %>%
    arrange(-abs(NES))

# Plot Cell-type enrichment
df_grcca_fgsea_cell %>%
    mutate(sig = case_when(padj < 0.001 ~ "***", padj < 0.01 ~ "**", padj < 0.05 ~ "*", TRUE ~ "")) %>%
    ggplot(aes(x = NES, y = reorder(cell_type, NES))) +
    geom_col(aes(fill = NES), color = "gray", shape = 21) +
    geom_text(aes(label = sig, x = NES), vjust = 0.75, size = 4, check_overlap = TRUE) +
    scale_fill_gradientn(colors = rev(brewer.rdbu(100)), limits = c(-3.39, 3.39)) +
    coord_cartesian(clip = "off") +
    labs(
        x = "NES", y = NULL,
        title = "Cell-type enrichment"
    ) +
    theme(legend.position = "none")
```

## GO GSEA

```{r go_gsea}
df_grcca_fgsea_go <- fgsea(
    pathways = go_pathways,
    stats = my_grcca_ranked,
    eps = 0
) %>%
    as_tibble() %>%
    dplyr::rename("go_id" = "pathway") %>%
    arrange(-abs(NES)) %>%
    left_join(df_go_terms) %>% # add names of GO terms
    dplyr::select(go_id, term, ontology, everything())
```

## Benchmark lists

```{r benchmark_gsea}
df_grcca_fgsea_benchmark <- fgsea(
    pathways = benchmarking_lists,
    stats = my_grcca_ranked,
    eps = 0
) %>%
    as_tibble() %>%
    dplyr::rename("benchmark_list" = "pathway") %>%
    arrange(-abs(NES))
```

## Export enrichment results as table

```{r export_results}
write_xlsx(
    list(
        "A.Cell-type" = df_grcca_fgsea_cell %>%
            mutate(genes = map_chr(leadingEdge, ~ f_collapse_genes(.x, index = 30000))) %>%
            dplyr::select(-leadingEdge) %>% unnest(cols = c(genes)),
        "B.GO" = df_grcca_fgsea_go %>%
            dplyr::filter(padj < 0.05) %>%
            mutate(genes = map_chr(leadingEdge, ~ f_collapse_genes(.x, index = 30000))) %>%
            dplyr::select(-leadingEdge) %>% unnest(cols = c(genes)),
        "C.Risk genes" = df_grcca_fgsea_benchmark %>%
            mutate(genes = map_chr(leadingEdge, ~ f_collapse_genes(.x, index = 30000))) %>%
            dplyr::select(-leadingEdge) %>% unnest(cols = c(genes))
    ),
    path = paste0(tables_dir, "SXX.GRCCA_transcripts_GSEA_enrichments.xlsx")
)
```

## Save objects for figures

```{r save}
save(df_grcca_fgsea_cell, df_grcca_fgsea_go, df_grcca_fgsea_benchmark,
    file = paste0(objects_dir, "GRCCA_transcripts_enrichment_benchmark_res.Rdata")
)
```
